Move code to util file; add 'few' notation for Arabic
