+++ app/webpack/taxa/show/components/charts.jsx
@@ -6,7 +6,7 @@ import c3 from "c3";
 import { schemeCategory10 } from "d3";
 import moment from "moment";
 import { Modal } from "react-bootstrap";
-import { objectToComparable } from "../../../shared/util";
+import { objectToComparable, compactDecimalFormattedNumber } from "../../../shared/util";
 
 class Charts extends React.Component {
   constructor( ) {
@@ -72,110 +72,7 @@ class Charts extends React.Component {
       return number.toExponential( 2 );
     }
     if ( number > 999 ) {
-      const compactNotations = Object.keys( I18n.t( "compact_number_formatting", { locale: "en" } ) );
-      let shortNumber;
-      let roundUp = false;
-
-      const lang = I18n.locale.split( "-" )[0];
-
-      const formatEastAsianLangs = ( ) => {
-        // special treatment for Chinese/Japanese/Korean
-        // where large numbers are grouped by four (instead of three)
-        // and start showing add'l digits from ten thousand (instead of one thousand)
-        // https://github.com/unicode-cldr/cldr-numbers-modern/blob/master/main/zh/numbers.json
-        let n;
-        if ( number < 1e9 ) {
-          n = number / 1e4;
-        } else if ( number < 1e13 ) {
-          n = number / 1e9;
-        } else {
-          n = number / 1e13;
-        }
-        return n;
-      };
-
-      const formatBasque = ( ) => {
-        // large numbers are grouped by six digits
-        // https://github.com/unicode-cldr/cldr-numbers-modern/blob/master/main/eu/numbers.json
-        let n;
-        if ( number < 1e6 ) {
-          n = number;
-        } else if ( number < 1e12 ) {
-          n = number / 1e6;
-        } else {
-          n = number / 1e12;
-        }
-        return n;
-      };
-
-      const formatGalician = ( ) => {
-        // thousands and billions don't have a symbol
-        // millions and trillions represented with same pattern as English ( 0 M, 00 M, 000 M)
-        // https://github.com/unicode-cldr/cldr-numbers-modern/blob/master/main/gl/numbers.json
-        let n;
-        if ( number < 1e6 ) {
-          n = number;
-        } else if ( number < 1e9 ) {
-          n = number / 1e6;
-        } else if ( number < 1e12 ) {
-          n = number;
-        } else {
-          n = number / 1e12;
-        }
-        return n;
-      };
-
-      const format = ( ) => {
-        // standard formatting for English and other Germanic/Romance languages
-        // https://github.com/unicode-cldr/cldr-numbers-modern/blob/master/main/en/numbers.json
-        let n;
-        if ( number < 1e6 ) {
-          if ( lang === "de" ) {
-            // German denotes 100,000 as 100.000 and starts compact notation from 1M onwards
-            // https://github.com/unicode-cldr/cldr-numbers-modern/blob/master/main/de/numbers.json
-            n = number;
-          } else {
-            n = number / 1e3;
-          }
-        } else if ( number < 1e9 ) {
-          n = number / 1e6;
-        } else if ( number < 1e12 ) {
-          n = number / 1e9;
-        } else {
-          n = number / 1e12;
-        }
-        return n;
-      };
-
-      // determine whether number needs to round up for proper formatting
-      // (i.e. from 1000K to 1M, 1000M to 1B)
-      const round = ( digits, length ) => shortNumber.toFixed( 0 ) === digits
-        && shortNumber < length;
-
-      if ( lang === "zh" || lang === "ja" || lang === "ko" ) {
-        shortNumber = formatEastAsianLangs( );
-        roundUp = round( 1e4, 1e13 );
-      } else if ( lang === "eu" ) {
-        shortNumber = formatBasque( );
-        roundUp = round( 1e6, 1e12 );
-      } else if ( lang === "gl" ) {
-        shortNumber = formatGalician( );
-        roundUp = shortNumber.toFixed( 0 ) === 1e3 && shortNumber === 1e9;
-      } else {
-        shortNumber = format( );
-        roundUp = round( 1e3, 1e12 );
-      }
-
-      const { length } = number.toString( );
-
-      const index = roundUp ? length - 3 : length - 4;
-      const notationIndex = Math.min( index, compactNotations.length - 1 );
-
-      return I18n.t(
-        `compact_number_formatting.${compactNotations[notationIndex]}`, {
-          count: roundUp ? 1 : I18n.toNumber( shortNumber, { precision: 0 } )
-        }
-      );
+      return compactDecimalFormattedNumber( number );
     }
     return I18n.toNumber( number, { precision } );
   }
